make Tuxedo work with the new `ChainSpecBuilder` pattern
