+++ tuxedo-core/Cargo.toml
@@ -33,6 +33,7 @@ sc-chain-spec = { optional = true, workspace = true }
 sc-client-api = { optional = true, workspace = true }
 sc-executor = { optional = true, workspace = true }
 sp-blockchain = { optional = true, workspace = true }
+sp-genesis-builder = { default-features = false, workspace = true }
 
 [dev-dependencies]
 array-bytes = { workspace = true }
@@ -47,6 +48,7 @@ std = [
 	"sp-std/std",
 	"serde/std",
 	"sp-api/std",
+	"sp-genesis-builder/std",
 	"sp-inherents/std",
 	"sp-io/std",
 	"sp-runtime/std",
