+++ tuxedo-template-runtime/src/lib.rs
@@ -9,7 +9,6 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
-#[cfg(feature = "std")]
 pub mod genesis;
 
 use parity_scale_codec::{Decode, Encode};
@@ -34,6 +33,7 @@ use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 
 use tuxedo_core::{
+    genesis::TuxedoGenesisConfigBuilder,
     tuxedo_constraint_checker, tuxedo_verifier,
     types::Transaction as TuxedoTransaction,
     verifier::{SigCheck, ThresholdMultiSignature, UpForGrabs},
@@ -182,8 +182,9 @@ pub enum OuterConstraintChecker {
     /// Checks that one winning claim came earlier than all the other claims, and thus
     /// the losing claims can be removed from storage.
     PoeDispute(poe::PoeDispute),
-    /// Set the block's timestamp via an inherent extrinsic.
-    SetTimestamp(timestamp::SetTimestamp<Runtime>),
+    // TODO restore this once inherent genesis can work with no_std
+    // /// Set the block's timestamp via an inherent extrinsic.
+    // SetTimestamp(timestamp::SetTimestamp<Runtime>),
     /// Upgrade the Wasm Runtime
     RuntimeUpgrade(runtime_upgrade::RuntimeUpgrade),
 
@@ -420,6 +421,19 @@ impl_runtime_apis! {
             _authority_id: sp_consensus_grandpa::AuthorityId,
         ) -> Option<sp_consensus_grandpa::OpaqueKeyOwnershipProof> {
             None
+        }
+    }
+
+    impl sp_genesis_builder::GenesisBuilder<Block> for Runtime {
+        fn create_default_config() -> Vec<u8> {
+            serde_json::to_vec(&genesis::development_genesis_transactions())
+                .expect("Development genesis transactions are valid.")
+        }
+
+        fn build_config(config: Vec<u8>) -> sp_genesis_builder::Result {
+            let genesis_transactions = serde_json::from_slice::<Vec<Transaction>>(config.as_slice())
+                .map_err(|_| "The input JSON is not a valid list of Transactions.")?;
+            TuxedoGenesisConfigBuilder::build(genesis_transactions)
         }
     }
